Moved the missing settings component into the components folder
