+++ client/src/app/components/add-rule/add-rule.component.ts
@@ -16,11 +16,11 @@ import { Component, OnInit, ViewChild } from '@angular/core';
 import { NgForm } from '@angular/forms';
 import { MatDialog } from '@angular/material/dialog';
 import { Router } from '@angular/router';
+import { MissingSettingsDialogComponent } from 'src/app/components/missing-settings-dialog/missing-settings-dialog.component';
 import { AgentSettingMetadata, ModelSpec } from 'src/app/interfaces/common';
 import { Comparator, Rule, RuleTargetAction } from 'src/app/interfaces/rule';
 import { SourceAgentMetadata } from 'src/app/interfaces/source';
 import { TargetAgentMetadata } from 'src/app/interfaces/target';
-import { MissingSettingsDialogComponent } from 'src/app/missing-settings-dialog/missing-settings-dialog.component';
 import { AgentsService } from 'src/app/services/agents.service';
 import { RulesService } from 'src/app/services/rules.service';
 import { UserService } from 'src/app/services/user.service';
