+++ client/src/app/components/missing-settings-dialog/missing-settings-dialog.component.scss
